feat: basic structure
