+++ backend/esg/urls.py
@@ -1,0 +1,7 @@
+# users/urls.py
+from django.urls import path
+from . import views
+
+urlpatterns = [
+    #path('', views.create_esg, name='create_esg_view')
+]
